+++ server/app.py
@@ -1,18 +1,14 @@
 from flask import Flask, jsonify, request, session
 from flask_sqlalchemy import SQLAlchemy
-# from routes import initRoute
-from routes.user import userBlueprint, jwt
-from routes.flashcard import flashcardBlueprint
-from routes.cardgroup import cardgroupBlueprint
+from blueprints.user.routes import userBlueprint, jwt
+from blueprints.flashcard.routes import flashcardBlueprint
+from blueprints.cardgroup.routes import cardgroupBlueprint
 from db import db
 from flask_cors import CORS
 from flask_jwt_extended import JWTManager, create_access_token, jwt_required, get_jwt_identity
 
-# db = SQLAlchemy()
-
 app = Flask(__name__)
 CORS(app, supports_credentials=True) # Support credentials to allow sessions in blueprints
-# CORS(app) # Support credentials to allow sessions in blueprints
 
 
 app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///mydb.db"
@@ -28,7 +24,6 @@ app.secret_key = 'super secret key'
 
 db.init_app(app)
 
-# app.register_blueprint(initRoute)
 app.register_blueprint(userBlueprint)
 app.register_blueprint(flashcardBlueprint)
 app.register_blueprint(cardgroupBlueprint)
